read extended
this is a leaked method that got 200k accounts banned. put this in your game or some shit and it'll basically do what it does i guess. the only way to use this safely is to blacklist yourself from the script by making it only run if the user isnt your user (which i don't know lua so i can't) or make an alt on a virtual machine (use a good vpn with the machine)